pyproject.toml
[tool.poetry]
name = "sandbox-sdk"
version = "0.1.0"
description = ""
authors = ["huang-jl <1046678590@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.7.3"
requests = "^2.32.3"
protobuf = "^5.27.1"
websockets = "^12.0"
jsonrpcclient = "^4.0.3"


[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.64.1"
pytest = "^8.2.2"
pytest-env = "^1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest_env]
SANDBOX_BACKEND_ADDR = {value = "192.168.47.247", skip_if_set = true}
SANDBOX_PROXY_PORT = {value = "6666", skip_if_set = true}
